(fix) use NSInteger instead, more clean

+++ SoObjects/Appointments/SOGoAppointmentFolder.m
@@ -1242,7 +1242,7 @@ firstInstanceCalendarDateRange: (NGCalendarDateRange *) fir
   iCalRepeatableEntityObject *component;
   iCalTimeZone *eventTimeZone;
   unsigned count, max;
-  int offset;
+  NSInteger offset;
   id tz;
 
   content = [theRecord objectForKey: @"c_cycleinfo"];
